Move writeToFile and dir methods into fileutils.py
