+++ minis/downloader/__init__.py
@@ -2,3 +2,4 @@ from .downloader import *
 from .show_info import *
 from .extract import *
 from .metadata import *
+from .fileutils import *
